#move-167 Refactored notification obligation enumeration into a seperate module to be used by both SR and KRR modules. Created a notification mapping to ensure that the query parameter is case insensitive Added parameter to be used with KRR. Remaining: logic in KRR to determine outfall of service record based on this requirement.
