+++ domain/src/main/java/no/difi/meldingsutveksling/NotificationObligation.java
@@ -1,4 +1,4 @@
-package no.difi.meldingsutveksling.serviceregistry.model;
+package no.difi.meldingsutveksling;
 
 public enum NotificationObligation {
     // if obligated then the sender is obligated to notify the recipient when sending a message
