+++ pom.xml
@@ -65,6 +65,7 @@
     <modules>
         <module>serviceregistry-server</module>
         <module>kontakt-reservasjonsregister-client</module>
+        <module>domain</module>
     </modules>
