+++ serviceregistry-server/pom.xml
@@ -255,6 +255,11 @@
             <groupId>no.difi.virksert</groupId>
             <artifactId>virksert-client</artifactId>
         </dependency>
+        <dependency>
+            <groupId>no.difi.meldingsutveksling</groupId>
+            <artifactId>service-registry-domain</artifactId>
+            <version>${project.parent.version}</version>
+        </dependency>
 
         <dependency>
             <groupId>org.springframework.boot</groupId>
